fix: Fixed filter by search
- Changed method `attributes`;
- Added `ilike` to query by filter;
