+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "unique-marketplace-api",
-  "version": "1.5.115",
+  "version": "1.5.116",
   "description": "Backend project for unique marketplace",
   "author": "Unique Network",
   "private": true,
